requirements.txt
backcall==0.1.0
dataclasses==0.6
decorator==4.4.0
Django==2.1.11
django-cors-headers==2.5.2
djangorestframework==3.9.2
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
ipython-genutils==0.2.0
jedi==0.13.3
mccabe==0.6.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pymongo==3.7.2
pytz==2018.9
six==1.12.0
sqlparse==0.2.4
traitlets==4.3.2
wcwidth==0.1.7
dj-database-url==0.5.0
psycopg2==2.8.3
django-extensions==2.2.5
Werkzeug==0.16.0
coreapi==2.3.3
PyYAML==5.1.2
drf-yasg==1.17.0
coverage==5.0.1
